+++ src/AngularSPAWebAPI/app/app.module.ts
@@ -1,5 +1,5 @@
 ﻿import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
+import { BrowserModule, Title } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { Http } from '@angular/http';
 
@@ -11,6 +11,7 @@ import { AuthenticationService } from './services/authentication.service';
 import { IdentityService } from './services/identity.service';
 
 import { AppComponent } from './app.component';
+import { HomeComponent } from './home/home.component';
 
 // angular2-jwt config for JiT and AoT compilation.
 import { AuthHttp, AuthConfig } from 'angular2-jwt';
@@ -31,11 +32,13 @@ export function getAuthHttp(http: Http) {
         SharedModule
     ],
     declarations: [
-        AppComponent
+        AppComponent,
+        HomeComponent
     ],
     exports: [
     ],
     providers: [
+        Title,
         AuthGuard,
         AuthenticationService,
         IdentityService,