+++ src/AngularSPAWebAPI/app/home/home.component.ts
@@ -1,6 +1,6 @@
 ﻿import { Component } from '@angular/core';
 
 @Component({
-    templateUrl: 'home.component.html'
+    templateUrl: './home.component.html'
 })
 export class HomeComponent { }