+++ src/AngularSPAWebAPI/app/resources/resources.component.ts
@@ -1,15 +1,17 @@
-﻿import { Component } from '@angular/core';
+﻿import { Component, OnInit } from '@angular/core';
 
 import { AuthHttp } from 'angular2-jwt';
 
 @Component({
-    templateUrl: 'resources.component.html'
+    templateUrl: './resources.component.html'
 })
-export class ResourcesComponent {
+export class ResourcesComponent implements OnInit {
 
     values: any;
 
-    constructor(private authHttp: AuthHttp) {
+    constructor(private authHttp: AuthHttp) { }
+
+    ngOnInit() {
         // Sends an authenticated request.
         this.authHttp.get("/api/values")
             .subscribe(